Add fit koopman operator params from data. The test does not really show a very satisfying long-term prediction.
To confirm:
1. Are there any constraints on the embedding function that can be used. The example used identity mapping as part of them so it can be easily retrieve the original one from lifted space.
While I remember some paper just threw NN to that and with a low-rank approximation. How much will this impact the performance?
The thing is, will the problem be ill-posed by finding all zero params when we optimize both NN and linear operator params? The responses did seem to have a tendency to return trivial flattened lines.

2. Using a random feature while not using pinverse did not repair things much.

3. If we have to use identity mapping as part, how can we achieve a low-rank approximation? Do an extra SVD?