Add a garage walker test to validate the code sanity of policy:
The reward indeed increases while the behaviour still looks crappy after 100 iterations. Not sure if this is normal or not.
More importantly, now all params, e.g. A, B and refs, are subject to policy gradient update. From the experience of koopman fitting, this seems not as effective as least square (but yeah, we dont have prediction auxiliary costs for now). So it is still unclear how could we embed least square estimation and see how it works. Ideally we could access the replybuffer and decide A and B on top of that rather than gradient descent but problem remains if an online SGD will work for that...